mangaplace/pyproject.toml
[project]
name = "mangaplace"
version = "0.1.0"
description = "A CLI tool to download manga."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.5",
    "pydantic>=2.8.2",
    "aiohttp>=3.10.5",
    "pillow>=10.4.0",
    "pylint>=3.2.7",
    "iterfzf>=1.4.0.54.3",
    "pyinstaller>=6.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
